+++ src/utils/etheroll-contract.js
@@ -1,5 +1,7 @@
-import getWeb3 from './get-web3';
 import etherollAbi from './etheroll-abi';
+
+// TODO require vs import
+const SolidityEvent = require('web3/lib/web3/event.js');
 
 const Networks = Object.freeze({ mainnet: 1, morden: 2, ropsten: 3 });
 
@@ -14,18 +16,142 @@ const etherscanUrls = {
 };
 
 
-function getEtherollContractSync(web3) {
-  const contractAddress = contractAddresses[web3.version.network];
-  return web3.eth.contract(etherollAbi).at(contractAddress);
-}
+class EtherollContract {
+  constructor(web3, address) {
+    this.web3 = web3;
+    this.address = address;
+    if (typeof address === 'undefined') {
+      this.address = contractAddresses[web3.version.network];
+    }
+    this.abi = etherollAbi;
+    this.web3Contract = web3.eth.contract(etherollAbi).at(this.address);
+  }
 
-function getEtherollContract() {
-  return getWeb3.then(results => (
-    getEtherollContractSync(results.web3)
-  ));
+  getSolidityEvents() {
+    const events = {};
+    this.abi.forEach((definition) => {
+      if (definition.type !== 'event') {
+        return;
+      }
+      events[definition.name] = new SolidityEvent(this.web3, definition, this.address);
+    });
+    return events;
+  }
+
+  // Returns sha3 signature of events, e.g.
+  // {'LogResult': '0x6883...5c88', 'LogBet': '0x1cb5...75c4'}
+  getEventSignatures() {
+    const signatures = {};
+    const events = this.getSolidityEvents();
+    Object.keys(events).forEach((eventName) => {
+      signatures[eventName] = events[eventName].signature();
+    });
+    return signatures;
+  }
+
+  getSolidityEvent(eventSignature) {
+    const events = this.getSolidityEvents();
+    const matchingEvent = Object.keys(events).filter(key => (
+      events[key].signature() === eventSignature.replace('0x', '')
+    ));
+    return events[matchingEvent];
+  }
+
+  decodeEvent(_evnt) {
+    // SolidityEvent.decode() seems to be mutating the object, hence the copy
+    const evnt = { ..._evnt };
+    const solidityEvent = this.getSolidityEvent(evnt.topics[0]);
+    const decoded = solidityEvent.decode(evnt);
+    return decoded;
+  }
+
+  // callback(error, result)
+  getTransactionLogs(callback) {
+    this.web3.eth.getBlockNumber((error, result) => {
+      if (error) {
+        console.log(error);
+      } else {
+        const { address } = this;
+        const lastBlock = result;
+        const fromBlock = lastBlock - 100;
+        const toBlock = lastBlock;
+        const options = {
+          address,
+          fromBlock,
+          toBlock,
+        };
+        const filter = this.web3.eth.filter(options);
+        filter.get(callback);
+      }
+    });
+  }
+
+  // callback(error, result)
+  watchTransactionLogs(callback) {
+    this.web3.eth.getBlockNumber((error, result) => {
+      if (error) {
+        console.log(error);
+      } else {
+        const { address } = this;
+        const lastBlock = result;
+        const fromBlock = lastBlock - 100;
+        const toBlock = lastBlock;
+        const options = {
+          address,
+          fromBlock,
+          toBlock,
+        };
+        const filter = this.web3.eth.filter(options);
+        filter.watch(callback);
+      }
+    });
+  }
+
+  // Merges bet logs (LogBet) with bet results logs (LogResult).
+  static mergeLogs(logBetEvents, logResultEvents) {
+    const mergedLogs = [];
+    let betId;
+    // per bet ID dictionary
+    const logResultEventsDict = {};
+    logResultEvents.forEach((logResultEvent) => {
+      betId = logResultEvent.args.BetID;
+      logResultEventsDict[betId] = logResultEvent;
+    });
+    let logResultEvent;
+    let mergedLog;
+    logBetEvents.forEach((logBetEvent) => {
+      betId = logBetEvent.args.BetID;
+      logResultEvent = logResultEventsDict[betId];
+      mergedLog = {
+        logBetEvent,
+        logResultEvent,
+      };
+      mergedLogs.push(mergedLog);
+    });
+    return mergedLogs;
+  }
+
+  // callback(error, result)
+  getMergedTransactionLogs(callback) {
+    this.getTransactionLogs((error, result) => {
+      if (error) {
+        console.log(error);
+      } else {
+        const decodedEvents = result.map(evnt => this.decodeEvent(evnt));
+        const logBetEvents = decodedEvents.filter(evnt => (
+          evnt.event === 'LogBet'
+        ));
+        const logResultEvents = decodedEvents.filter(evnt => (
+          evnt.event === 'LogResult'
+        ));
+        const mergedLogs = EtherollContract.mergeLogs(logBetEvents, logResultEvents);
+        callback(error, mergedLogs);
+      }
+    });
+  }
 }
 
 
 export {
-  getEtherollContract, getEtherollContractSync, etherscanUrls, Networks, contractAddresses,
+  EtherollContract, etherscanUrls, Networks, contractAddresses,
 };