+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Provide Docker image, refs #13
   - CI testing with Travis, refs #7
   - Show error message on missing web3-wallet, refs #4 and #5
+  - Show transaction history, refs #1
 
 ## [v20181016]